refactor(itest): Bundle Custom Expectations
This provides a better overview, and most likely fixes issues,
for matchers having the same name.
