+++ itest/src/ContentLink.test.ts
@@ -1,5 +1,4 @@
-import "./expect/ElementHandleExpectations";
-import "./expect/LocatorExpectations";
+import "./expect/Expectations";
 import { ApplicationWrapper } from "./aut/ApplicationWrapper";
 import { contentUriPath } from "@coremedia/ckeditor5-coremedia-studio-integration/content/UriPath";
 import { a, p, richtext } from "@coremedia-internal/ckeditor5-coremedia-example-data/RichTextBase";
