+++ itest/src/expect/ElementHandleExpectations.ts
@@ -24,20 +24,3 @@ export interface ElementHandleMatchers<R = unknown, T = unknown> {
     ? (expected: string) => R
     : "Type-level Error: Received value must be an ElementHandle";
 }
-
-/**
- * Tell TypeScript to know of new matchers.
- */
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace jest {
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    interface Expect extends ElementHandleMatchers {}
-
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface,@typescript-eslint/ban-types
-    interface Matchers<R = unknown, T = {}> extends ElementHandleMatchers<R, T> {}
-
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    interface InverseAsymmetricMatchers extends ElementHandleMatchers {}
-  }
-}
